+++ src/functor.jl
@@ -285,8 +285,9 @@ _paramtype(::Type{T}, x::AbstractArray{<:Complex{<:AbstractFloat}}) where {T<:Ab
 """
     f32(m)
 
-Converts the `eltype` of model's parameters to `Float32` (which is Flux's default).
-Recurses into structs marked with [`@functor`](@ref). Preserves integer arrays.
+Converts the `eltype` of model's *floating point* parameters to `Float32` (which is Flux's default).
+Recurses into structs marked with [`@functor`](@ref).
+
 See also [`f64`](@ref) and [`f16`](@ref).
 """
 f32(m) = _paramtype(Float32, m)
@@ -294,8 +295,9 @@ f32(m) = _paramtype(Float32, m)
 """
     f64(m)
 
-Converts the `eltype` of model's parameters to `Float64`.
-Recurses into structs marked with [`@functor`](@ref). Preserves integer arrays.
+Converts the `eltype` of model's *floating point* parameters to `Float64`.
+Recurses into structs marked with [`@functor`](@ref).
+
 See also [`f32`](@ref) and [`f16`](@ref).
 """
 f64(m) = _paramtype(Float64, m)
@@ -303,11 +305,13 @@ f64(m) = _paramtype(Float64, m)
 """
     f16(m)
 
-Converts the `eltype` of model's parameters to `Float16`.
-Recurses into structs marked with [`@functor`](@ref). Preserves integer arrays.
+Converts the `eltype` of model's *floating point* parameters to `Float16`.
+Recurses into structs marked with [`@functor`](@ref).
 
 Support for `Float16` is limited on many CPUs. Julia may
 convert to `Float32` for each operation, which is slow.
+
+See also [`f32`](@ref) and [`f64`](@ref).
 
 # Example
 ```jldoctest