+++ NEWS.md
@@ -1,8 +1,8 @@
 # Flux Release Notes
 
 ## v0.13.16
-* `f16, f32, f64` now preserve integer arrays.
-* `f16, f32, f64` can now handle `Complex` arrays.
+* `f16, f32, f64` now specifically target floating point arrays (i.e. integers arrays and other types are preserved).
+* `f16, f32, f64` can now handle `Complex{<:AbstractFloat}` arrays.
 
 ## v0.13.15
 * Added [MultiHeadAttention](https://github.com/FluxML/Flux.jl/pull/2146) layer.